chore: Add package to handle all rendering
This new package merges the functionality from the Layer and Item
packages and simplifies their combined behaviour dramatically.

A new canvas type  has been added, these are called Scenes, and handle
all of the grouping that Layers used to handle. Layers are now much
simpler, and remove the need for the Items class as they now also handle
the drawing commands as well.